backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pdfplumber
python-docx
python-dotenv